services/earnings-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
httpx==0.27.0
pydantic==2.7.1
prometheus-client==0.20.0
python-dotenv==1.0.1
aiofiles==23.2.1
